backend/requirements.txt
# ATOM v2 Python Dependencies

# Core dependencies
aiohttp>=3.8.0
websockets>=11.0.0
python-dotenv>=1.0.0

# Blockchain and Web3
web3>=5.31.0
eth-account>=0.9.0
eth-utils>=2.0.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0

# HTTP and API
requests>=2.28.0
httpx>=0.24.0

# Cryptography
cryptography>=41.0.0
pycryptodome>=3.18.0

# Logging and monitoring
structlog>=23.0.0
prometheus-client>=0.17.0

# Database (optional)
sqlalchemy>=2.0.0
alembic>=1.11.0
psycopg2-binary>=2.9.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-mock>=3.11.0

# Development tools
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0

# Configuration and validation
pydantic>=2.0.0
jsonschema>=4.19.0

# Utilities
click>=8.1.0
rich>=13.0.0
typer>=0.9.0
